Remove duplicate consecutive hops in a prior step.
