+++ analysis/README.md
@@ -27,6 +27,7 @@ This produces a file that contains one route on each line, for each source IP, a
 
 - We next convert each IP address to a (lat, long) geocoordinate using the ITDK `.nodes.geo` database:
 ```Shell
+# (optionally, remove duplicate consecutive hops) --remove-duplicate-consecutive-hops
 ./itdk_geo.py --convert-ip-to-latlon --routes_file routes.aws.us-west-1.us-east-1.by_ip 1> routes.aws.us-west-1.us-east-1.by_geo
 ```
 
@@ -49,8 +50,8 @@ awk -F '\t' '{print $1}' routes.aws.us-west-1.us-east-1.by_geo.physical > routes
 
 - Finally, we can export the distribution of geo-coordinates or ISOs for easy lookup later (e.g. in a database).
 ```Shell
-# (optionally, include additional metrics and remove duplicate consecutive hops) --include hop_count distance_km --remove-duplicate-consecutive-hops
-# (optionally, include iGDB route fiber information using earlier files) --include fiber_wkt_paths fiber_types --physical-routes-tsv routes.aws.us-west-1.us-east-1.by_geo.physical
+# (optionally, include additional metrics) --include hop_count distance_km
+# (optionally, further include iGDB route fiber information using earlier files) --include fiber_wkt_paths fiber_types --physical-routes-tsv routes.aws.us-west-1.us-east-1.by_geo.physical
 ./distribution.routes.py --export-routes-distribution --routes_file routes.aws.us-west-1.us-east-1.by_geo > routes.aws.us-west-1.us-east-1.by_geo.distribution
 ./distribution.routes.py --export-routes-distribution --routes_file routes.aws.us-west-1.us-east-1.by_iso > routes.aws.us-west-1.us-east-1.by_iso.distribution
 ```
